+++ packages/ui/index.ts
@@ -4,6 +4,7 @@ export * from "./components/Button";
 export * from "./components/ButtonGroup";
 export * from "./components/BigText";
 export * from "./components/ProgressBar";
+export * from "./components/Input";
 export * from "./components/Text";
 export * from "./components/Scrollbar"
 export * from "./components/WindowProvider";