+++ apps/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.3.2
+
+### Patch Changes
+
+- input component added
+- Updated dependencies
+  - @duck4i/retro-ui@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes