+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @duck4i/retro-ui
 
+## 0.3.2
+
+### Patch Changes
+
+- input component added
+
 ## 0.3.1
 
 ### Patch Changes